Update extension.ts
Since "vscode.workspace.rootPath" has been deprecated, "vscode.workspace.workspaceFolders" should be used. However, due to the multi-root workspaces of VSCode, it would be more prudent to identify the current active file (being open in the VSCode editor), and look for it's root folder amongst the others. This is now implemented with many comments.